+++ frame_forge.py
@@ -159,6 +159,18 @@ if __name__ == "__main__":
         help="Subtitle border style (0=off, 1=outline, 3=opaque box) [defaults to '0']",
     )
     parser.add_argument(
+        "--sub-outline-width",
+        type=int,
+        default=1,
+        help="Subtitle outline width (defaults to '1')",
+    )
+    parser.add_argument(
+        "--sub-shadow-depth",
+        type=int,
+        default=0,
+        help="Subtitle shadow depth (defaults to '0')",
+    )
+    parser.add_argument(
         "--source-sub-title",
         type=str,
         default="Source",
@@ -257,6 +269,8 @@ if __name__ == "__main__":
             sub_scale_y=args.sub_scale_y,
             sub_spacing=args.sub_spacing,
             sub_border_style=args.sub_border_style,
+            sub_outline_width=args.sub_outline_width,
+            sub_shadow_depth=args.sub_shadow_depth,
             source_sub_title=args.source_sub_title,
             release_sub_title=args.encode_sub_title,
         )