+++ frame_forge/__init__.py
@@ -50,6 +50,8 @@ class GenerateImages:
         sub_scale_y: int,
         sub_spacing: int,
         sub_border_style: int,
+        sub_outline_width: int,
+        sub_shadow_depth: int,
         source_sub_title: str,
         release_sub_title: str,
     ):
@@ -91,6 +93,8 @@ class GenerateImages:
         self.sub_scale_y = sub_scale_y
         self.sub_spacing = sub_spacing
         self.sub_border_style = sub_border_style
+        self.sub_outline_width = sub_outline_width
+        self.sub_shadow_depth = sub_shadow_depth
         self.source_sub_title = source_sub_title
         self.release_sub_title = release_sub_title
 
@@ -137,7 +141,7 @@ class GenerateImages:
             f"{self.sub_font_name},{self.sub_size},{color},{secondary_color},{outline_color},{back_color},"
             f"{self.sub_bold},{self.sub_italic},{self.sub_underline},{self.sub_strikeout},"
             f"{self.sub_scale_x},{self.sub_scale_y},{self.sub_spacing},0,{self.sub_border_style},"
-            f"1,0,{self.sub_alignment},10,10,10,1"
+            f"{self.sub_outline_width},{self.sub_shadow_depth},{self.sub_alignment},10,10,10,1"
         )
         sync_sub_base = (
             "Segoe UI,{size},&H31FF31&,&H00000000,&H00000000,&H00000000,"